回退 'Pull Request !191 : Support play audio/video on L2 device and other modifications.'

+++ engine/plugin/plugins/ffmpeg_adapter/video_encoder/video_ffmpeg_encoder_plugin.h
@@ -117,7 +117,7 @@ private:
 
     uint32_t width_ {0};
     uint32_t height_ {0};
-    uint32_t frameRate_ {0};
+    uint64_t frameRate_ {0};
     VideoPixelFormat pixelFormat_ {};
 
     mutable OSAL::Mutex avMutex_ {};
